SFEperl-log-log4perl: package for perl-522 is added, for perl-520 is obsolete
--HG--
branch : spec_SFEperl-log-log4perl

+++ specs/SFEperl-log-log4perl.spec
@@ -4,7 +4,7 @@
 %define build510 %( if [ -x /usr/perl5/5.10/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build512 %( if [ -x /usr/perl5/5.12/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define build516 %( if [ -x /usr/perl5/5.16/bin/perl ]; then echo '1'; else echo '0'; fi)
-%define build520 %( if [ -x /usr/perl5/5.20/bin/perl ]; then echo '1'; else echo '0'; fi)
+%define build522 %( if [ -x /usr/perl5/5.22/bin/perl ]; then echo '1'; else echo '0'; fi)
 %define include_executable 1
 
 %define cpan_name Log-Log4perl
@@ -14,8 +14,8 @@
 Summary:               Log4j implementation for Perl
 Name:                  SFEperl-%{sfe_cpan_name}
 IPS_package_name:      library/perl-5/%{ips_cpan_name}
-Version:               1.46
-IPS_component_version: 1.46
+Version:               1.49
+IPS_component_version: 1.49
 License:               unknown
 URL:                   https://metacpan.org/pod/Log::Log4perl
 Source0:               http://cpan.metacpan.org/authors/id/M/MS/MSCHILLI/Log-Log4perl-%{version}.tar.gz
@@ -30,7 +30,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-584
 Summary:          Log4j implementation for Perl
 BuildRequires:    runtime/perl-584 = *
 BuildRequires:    library/perl-5/extutils-makemaker-584
+BuildRequires:    library/perl-5/file-path-584
+BuildRequires:    library/perl-5/pathtools-584
+BuildRequires:    library/perl-5/test-simple-584
 Requires:         runtime/perl-584 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/file-path-584
 Requires:         library/perl-5/pathtools-584
 Requires:         library/perl-5/test-simple-584
@@ -45,7 +49,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-510
 Summary:          Log4j implementation for Perl
 BuildRequires:    runtime/perl-510 = *
 BuildRequires:    library/perl-5/extutils-makemaker-510
+BuildRequires:    library/perl-5/file-path-510
+BuildRequires:    library/perl-5/pathtools-510
+BuildRequires:    library/perl-5/test-simple-510
 Requires:         runtime/perl-510 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/file-path-510
 Requires:         library/perl-5/pathtools-510
 Requires:         library/perl-5/test-simple-510
@@ -60,7 +68,11 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-512
 Summary:          Log4j implementation for Perl
 BuildRequires:    runtime/perl-512 = *
 BuildRequires:    library/perl-5/extutils-makemaker-512
+BuildRequires:    library/perl-5/file-path-512
+BuildRequires:    library/perl-5/pathtools-512
+BuildRequires:    library/perl-5/test-simple-512
 Requires:         runtime/perl-512 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/file-path-512
 Requires:         library/perl-5/pathtools-512
 Requires:         library/perl-5/test-simple-512
@@ -75,7 +87,12 @@ IPS_package_name: library/perl-5/%{ips_cpan_name}-516
 Summary:          Log4j implementation for Perl
 BuildRequires:    runtime/perl-516 = *
 BuildRequires:    library/perl-5/extutils-makemaker-516
+Requires:         library/perl-5/%{ips_cpan_name}
+BuildRequires:    library/perl-5/file-path-516
+BuildRequires:    library/perl-5/pathtools-516
+BuildRequires:    library/perl-5/test-simple-516
 Requires:         runtime/perl-516 = *
+Requires:         library/perl-5/%{ips_cpan_name}
 Requires:         library/perl-5/file-path-516
 Requires:         library/perl-5/pathtools-516
 Requires:         library/perl-5/test-simple-516
@@ -84,25 +101,29 @@ Requires:         library/perl-5/test-simple-516
 Log4j implementation for Perl
 %endif
 
-%if %{build520}
-%package 520
-IPS_package_name: library/perl-5/%{ips_cpan_name}-520
+%if %{build522}
+%package 522
+IPS_package_name: library/perl-5/%{ips_cpan_name}-522
 Summary:          Log4j implementation for Perl
-BuildRequires:    runtime/perl-520 = *
-BuildRequires:    library/perl-5/extutils-makemaker-520
-Requires:         runtime/perl-520 = *
-Requires:         library/perl-5/file-path-520
-Requires:         library/perl-5/pathtools-520
-Requires:         library/perl-5/test-simple-520
+BuildRequires:    runtime/perl-522 = *
+BuildRequires:    library/perl-5/extutils-makemaker-522
+BuildRequires:    library/perl-5/file-path-522
+BuildRequires:    library/perl-5/pathtools-522
+BuildRequires:    library/perl-5/test-simple-522
+Requires:         runtime/perl-522 = *
+Requires:         library/perl-5/%{ips_cpan_name}
+Requires:         library/perl-5/file-path-522
+Requires:         library/perl-5/pathtools-522
+Requires:         library/perl-5/test-simple-522
 
-%description 520
+%description 522
 Log4j implementation for Perl
 %endif
 
 
 %prep
 %setup -q -n %{cpan_name}-%{version}
-rm -rf %{buildroot}
+[ -d %{buildroot} ] && rm -rf %{buildroot}
 
 %build
 build_with_makefile.pl_for() {
@@ -115,8 +136,17 @@ build_with_makefile.pl_for() {
     ${bindir}/perl Makefile.PL PREFIX=%{_prefix} \
                    DESTDIR=$RPM_BUILD_ROOT \
                    LIB=${vendor_dir}
-    make
-    [ ${test} = 'without_test' ] || make test
+
+    echo ${perl_ver} | egrep '5\.(84|12)' > /dev/null
+    if [ $? -eq 0 ]
+    then
+        make CC='cc -m32' LD='cc -m32'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m32' LD='cc -m32'
+    else
+        make CC='cc -m64' LD='cc -m64'
+        [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || make test CC='cc -m64' LD='cc -m64'
+    fi
+
     make pure_install
 }
 
@@ -131,17 +161,47 @@ build_with_build.pl_for() {
                    --installdirs vendor \
                    --destdir $RPM_BUILD_ROOT
     ${bindir}/perl ./Build
-    [ ${test} = 'without_test' ] || ${bindir}/perl ./Build test
+    [ "x${PERL_DISABLE_TEST}" = 'xtrue' ] || [ "x${test}" = 'xwithout_test' ] || ${bindir}/perl ./Build test
     ${bindir}/perl ./Build install --destdir $RPM_BUILD_ROOT
+    ${bindir}/perl ./Build clean
 }
 
 modify_bin_dir() {
-  perl_ver=$1
-  if [ -d $RPM_BUILD_ROOT/usr/bin ]
-  then
-    [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
-    mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
-  fi
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/bin ]
+    then
+      [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver} ] || mkdir -p $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+      mv $RPM_BUILD_ROOT/usr/bin $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin
+    fi
+      
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin ]
+    then
+        for i in $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/bin/*
+        do
+            sed -i.bak -e "s/\/usr\/bin\/env ruby/\/usr\/perl5\/${perl-ver}\/bin\/ruby/" ${i}
+            [ -f ${i}.bak] || rm -f ${i}.bak
+        done
+    fi
+}
+
+modify_man_dir() {
+    perl_ver=$1
+    if [ -d $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man ]
+    then
+        if [ -d $RPM_BUILD_ROOT%{_datadir}/man ]
+        then
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        else
+            mkdir -p $RPM_BUILD_ROOT%{_datadir}
+            mv $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man $RPM_BUILD_ROOT%{_datadir}/
+            rm -rf $RPM_BUILD_ROOT/usr/perl5/${perl_ver}/man
+        fi
+        if [ %{include_executable} -eq 0 ]
+        then
+            rmdir $RPM_BUILD_ROOT/usr/perl5/${perl_ver}
+        fi
+
+    fi
 }
 
 build_for() {
@@ -153,7 +213,8 @@ build_for() {
     build_with_makefile.pl_for $*
   fi
 
-    modify_bin_dir $*
+  modify_bin_dir $*
+  modify_man_dir $*
 }
 
 # To build without test, pass 'without_test' to build_for commaond.
@@ -174,14 +235,14 @@ build_for 5.12
 build_for 5.16
 %endif
 
-%if %{build520}
-build_for 5.20
+%if %{build522}
+build_for 5.22
 %endif
 
 %install
-mkdir -p $RPM_BUILD_ROOT%{_datadir}
 if [ -d $RPM_BUILD_ROOT%{_prefix}/man ]
 then
+    mkdir -p $RPM_BUILD_ROOT%{_datadir}
     mv $RPM_BUILD_ROOT%{_prefix}/man $RPM_BUILD_ROOT%{_datadir}
 fi
 if [ -d $RPM_BUILD_ROOT%{_datadir}/man/man3 ]
@@ -236,18 +297,19 @@ rm -rf %{buildroot}
 %endif
 %endif
 
-%if %{build520}
-%files 520
+%if %{build522}
+%files 522
 %defattr(0755,root,bin,-)
 %dir %attr (0755, root, sys) /usr
-/usr/perl5/vendor_perl/5.20
+/usr/perl5/vendor_perl/5.22
 %if %{include_executable}
-/usr/perl5/5.20
+/usr/perl5/5.22
 %endif
 %endif
-
 
 %changelog
+* Tue Apr 04 2017 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
+- package for perl-522 is added, for perl-520 is obsolete
 * Thu Nov 12 2015 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
 - bump to 1.46 and build packages for perl-510, perl-516 and perl-520
 * Thu Nov 13 2013 - Fumihisa TONAKA <fumi.ftnk@gmail.com>
